pyproject.toml
[tool.poetry]
name = "py-voice-memo-export"
version = "0.1.0"
description = "A Python tool for exporting Apple Voice Memos effortlessly. PyVoiceMemoExport provides a simple solution to back up your voice memos."
authors = ["Robin Schulz <hello@rschu.me>"]
license = "MIT"
repository = "https://github.com/bulletinmybeard/py-voice-memo-export"
homepage = "https://rschu.me/"
readme = "README.md"
packages = [
    { include = "pyvme" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.4.0"
python-dateutil = "^2.9.0.post0"
colorama = "^0.4.6"
psutil = "^6.0.0"
jinja2 = "^3.1.2"
python-slugify = "^8.0.4"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.11.2"
flake8 = "^7.1.1"
pytest-cov = "^5.0.0"
pre-commit = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pyvme = "pyvme.__main__:main"
